+++ R/date.R
@@ -978,18 +978,6 @@ date_format.Date <- function(x,
 
 # ------------------------------------------------------------------------------
 
-#' @export
-date_zone.Date <- function(x) {
-  abort("Can't get the zone of a 'Date'.")
-}
-
-#' @export
-date_set_zone.Date <- function(x, zone) {
-  abort("Can't set the zone of a 'Date'.")
-}
-
-# ------------------------------------------------------------------------------
-
 #' Parsing: date
 #'
 #' @description
@@ -1978,6 +1966,10 @@ date_count_between_impl <- function(start,
 }
 
 # ------------------------------------------------------------------------------
+
+is_date <- function(x) {
+  inherits(x, "Date")
+}
 
 check_date <- function(x,
                        ...,